requirements.txt
fastapi
uvicorn
openai
python-dotenv
PyPDF2
python-docx
python-multipart
openai
python-multipart
chromadb
langchain
langchain-openai
python-docx
tiktoken
numpy
